src refactored into 3 files
