+++ test/main.js
@@ -49,6 +49,9 @@ describe('app', (suite) => {
 
   it('should exist', (done) => {
     should.exist(g.app)
+    should.exist(k.list_query)
+    should.exist(k.list_middleware)
+    should.exist(k.create_relation_middleware)
     done()
   })
